pyproject.toml
[project]
name = "agno-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.5.17",
    "ddgs>=9.14.0",
    "fastapi>=0.136.0",
    "googlesearch-python>=1.3.0",
    "openai>=2.32.0",
    "pycountry>=26.2.16",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]
